backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.1
groq==0.11.0
# Pin httpx for groq client compatibility (proxies kw only supported <0.28)
httpx==0.27.2
python-dotenv==1.0.1

